Fixed some FindBugs errors  - Removed some superfluous method invocations  - Fixed an infinite recursion but in ActionManager  - Swapped some string concatenation for StringBuilders in ConditionTree  - Fixed the regex in VLC media source when running under windows  - Removed some unused/needed variables/code  - Renamed CategoryChangeListener methods to follow naming conventions  - Fixed useless string equals checks in the windows installer  - Fixed some genericity errors in ErrorManager and ButtonBar  - Fixed some crap primitive coercion in DurationEditor  - toString in InviteLabel is now useful  - Fixed some potential NPEs in the equals method in Profile  - Fixed some casting errors in JWrappingLabel  - Changed the scope of some fields in CipherUtils
git-svn-id: http://svn.dmdirc.com/trunk@3629 00569f92-eb28-0410-84fd-f71c24880f43

+++ src/com/dmdirc/logger/ErrorManager.java
@@ -146,7 +146,7 @@ public final class ErrorManager implements Serializable, Runnable {
      *
      * @return ProgramError with specified ID
      */
-    public ProgramError getError(final int id) {
+    public ProgramError getError(final long id) {
         return errors.get(id);
     }
